Added support for building using ESP-IDF 4.4.4, fixed wic64.h
Restructured to allow building with ESP-IDF toolchain. We need to use
a 4.4.x version because the arduino-esp32 library has not yet been
ported to the current 5.x version of ESP-IDF.

Added required components as submodules:

- components/Adafruit-GFX-Library
- components/Adafruit_BusIO
- components/Adafruit_SSD1306
- components/arduino (arduino-esp32)
- components/OneButton

The arduino component has been checked out to version 2.0.7 to make
sure both toolchains use the exact same version of the arduino-esp32
library. For unknown reasons, the directory needs to be named
"arduino" in order for the ESP-IDF toolchain to resolve this
component properly.

components/OneButtonComponent contains the CMakeLists.txt required for
ESP-IDF to resolve the OneButton component -- upstream lacks this file.

Added wic64.ino, this is just a stub required because the Arduino IDE
expects an .ino file to be present, even if empty. Note that this file
also needs to have the same basename as the directory in which it
resides.

Added missing function declarations in wic64.h -- The Arduino IDE
automagically adds these in a preprocessing step if the are missing.