Started reimplementing the firmware using a component based approach
After some inspection, I have found the previous implementation to be
unmaintainable. It makes more sense to start over using a component based
approach with proper separation of concerns assuring both extensibility
and maintainability in the future. The existing features will be gradually
reimplemented once a suitable framework is in place.

Work has started on the following components:

- Connection

Responsible for setting up the WiFi connection, handling WiFi events,
initiating reconnects on loss of connection and updating the display.
This component is already functional.

- Display

Abstracts the display behind a simple facade allowing the user to set
the individual values. This component is already functional as well,
displaying the IP, the SSID, status and RSSI for now.

- Userport

Abstracts the low level communication protocol between the host and the
esp32 using the userport. Responsible for managing all the technical
details required for transfers. Partially implemented.

The following components will be implemented in the future:

- Service

Main component. Implements the high level communication, accepting
requests from the host, executing the requested actions and sending
the appropriate responses.

- Buttons & Leds

Rotating the display, detaching the device from the userport, previous
Status display using the on board leds.

- Webserver

Local web interface for doing firmware updates and WiFi configuration
(SSID & password).

The previous implementation has been moved to the legacy directory.