Userport: Added isReadyToSend/Receive(), improved comments, readability
Renamed isReady() to isReadyToReceive(), added isReadyToSend(). If the
userport is connected, it's ready to receive when the data direction line
PA2 is high, and ready to send when it's low.

The handshake line form the esp to the c64 (FLAG2) is  now explicitly
initialized to high in Userport::connect().

To improve readability, synomyms for the control signal lines PC2, PA2
and FLAG2 have been added and macros for reading and writing GPIO pins
have been added and applied.