Userport: Implemented send and receive transfers with async callbacks
The methods send() and receive() can be used to perform the low level
transfers between the C64 and the ESP. Since communication is entirely
driven by interrupts, transfers run asynchronously and call the provided
onSuccess() callback once the transfer is complete.

The ISRs onHandshakeSignalReceived() and onDataDirectionChanged() handle
the control signals from the C64 (PC2 for handshakes, PA2 for data
direction). The data direction interrupt is always enabled as long as the
port is connected, while the handshake interrupt is only enabled during
transfers.

Internally the enums TRANSFER_TYPE and TRANSFER_STATE are used to
distinguish between transfer types and to manage transfer states.

Receiving is straightforward since the C64 initiates the transfer,
so they always start in TRANSFER_STATE_RUNNING.

When sending, the C64 first sets its port to input and then requests a
change to output from the ESP, waits for an additional handshake to
confirm the change of direction and only then starts to read data from
the port. Thus sending transfers start in TRANSFER_STATE_PENDING and
change to TRANSFER_STATE_RUNNING once the data direction has been
successfully negotiated using the mechanism described above.

Thus sending transfers start in TRANSFER_STATE_PENDING. If a transfer is
found to be pending when receiving a request to change the data direction
to output, the direction is first changed and the acknowledged with an
additional handshake, and the state of the transfer changes to
TRANSFER_STATE_RUNNING.

The methods isIdle(), isTransferPending(), isSending() and isReceiving()
have been added to make the code more readable. In addition, synonmyms
for the control Signals have been added for the same reason.

Detection and handling of timeouts are still missing.