Userport: Implemented optional transfer timeouts using a FreeRTOS task
The send() and receive() methods have been overloaded to accept an
additional timeout argument that allows the user to specify a transfer
timeout in milliseconds. If no timeout argument is supplied, the timeout
defaults to 2000ms (same as in legacy firmware). If a timeout is supplied,
the transfer will be aborted if no activity from the C64 has occured for
more than timeout milliseconds. If the constant Userport::NO_TIMEOUT is
supplied, the transfer will never time out. This will be used to wait
indefinitely for the C64 to initialize the next transfer.

The timeout functionality has been implemented using a FreeRTOS task that
checks the time since last activity every 100ms during transfers. This task
is created when a transfer is started with a timeout and it is deleted
again if the transfer completes or has timed out. When a timeout occurs,
the current transfer is aborted. This means the port lines are set back to
input, the handshake interrupt is deinstalled and the transfer type and
state are reset. In other words, the userport is reset to its idle state.

Activity here means that either a handshake signal has been received or a
change of direction has been requested by the C64.