Userport/Service: Refactored Userport, started implementing Service
Changed the way that initiation of a transfer session from the C64 is
handled since the previous implementation worked for sessions that were
initiated with some amount of delay after the previous session was done,
but failed for sessions immediately initiated after a previous session.

This was due to the fact that the transfer code was still handling the
final send transfer callback when the next initial magic byte containing
the API version was send from the C64, thus missing the start of the next
session. For this reason the entire session is now handled inside its own
Session Task so that the next session trigger is registered even though the
last session has not quite finished yet.

The session begins when a handshake signal is received when the current
transfer type is TRANSFER_TYPE_NONE (this is the case even when the previous
session task is still busy whith the final success callback). Then a new
session task is created which simply calls Userport::acceptRequest() and
then (after the request has been handled) deletes itself again.

Userpor::acceptRequest() will just read the first byte of the incoming
request, which is usually the magic byte $57 ('W'), e.g. the current
API id. It then asks the Service component if the received byte denotes
a supported API by calling Service::supports(uint8_t api_id). If an
unsupported magic byte has  been received, the transfer will be aborted
with an appropriate error message. Else further processing of the
request is delegated to Service::receiveRequest(uint8_t api_id).

Service::receiveRequest() will then start a partial receive transfer for
the next 3 bytes which constitute the command header, delegating to
Service::parseRequestHeaderV1. This method will then parse the command id
and the length of the single argument, create and prepare an instance of
Service::Command and then start the next full receive transfer for the
argument data, delegating to Service::postRequestReceivedEvent().

Up to now, all receiving transfers have run from the ISR contexts.
The Service component now runs its own event loop task in order to handle
fully reveived requests outside of the ISR context, since lengthy actions
(such as fetching data from the net) can not be performed in ISRs.

So the onSuccess callback of the last receive transfer just posts the
corresponding event to the Service event loop via
Service::postRequestReceivedEvent(). This event will in turn
be handled in Service::onRequestReceived(), now in the context of the
Service event loop task rather than the ISR context.

Currently onRequestReceived() simply logs info about the received commands
and then sends a hardcoded response which is specific to some simple test
code still outside of this repository. It will contain the logic to
dispatch the request to execute the specified command in the future.

The Service methods supports(api_id) and receiveRequest(api_id) allow the
implementation of a more flexible API in the future without the need to
immediately drop the legacy API. The plan here is to be able to handle
different request header formats in the future without having to implement
different command executions, which is why the Service::Command object
already supports the notion of more than one argument. In other words,
the parsing of the legacy and future binary command formats should result
in the same Service::Command objects, which can then be handled transparently
by the same actual command implementations.

The Userport code has also been cleaned up and some unused methods have been
removed. The timeout on the ESP side has been reduced to 10ms instead of
the previous 2000ms since it can be assumed that when the C64 shows no
activity for 10ms during a running transfer something has already gone
wrong on that side.

In addition the timeout argument has been removed from the send/receive
methods since it is no longer needed to specify no timeout. The original
idea was to just always start a new receive transfer without timeout to
wait for the next session to be initiated, but this is no longer necessary
since transfers now only start once the initial handshake has been received
and a new transfer session has been started.