Utilities: Refactored hexdump() to log_v() every line, added vTaskDelay()
Previously the whole string has been allocated in advance and then logged
at once. While this is more readable, it wastes memory and blocks the
running task too long for larger amounts of data. Now only one line is
allocated and logged at a time, and a vTaskDelay() is done after each
line. The latter prevents the task watchdog of this and other tasks from
triggering when larger amounts of data are dumped.

In addition, the log level has been increased to "verbose" and the
hexdump() function exits right away if the log level is lower.