Userport: Refactored to use events instead of doing work in ISRs
Previously all transfers where still done from ISR context, but I have
now learned that this is considered bad practice. In particular I ran
into issues when logging larger amounts of data using the hexdump()
function from ISR context, since this would eventually trigger the ISR
watchdog.

Instead the Userport now has its own event loop task. The ISRs for the
handshake signal and data direction lines now simply post the
appropriate events to the Userport event loop, so that the bulk of the
work is always done in this task and the hardware interrupts are as
short as possible.

Using an event loop also has the benefit of avoiding missed interrupts
since the events are queued in the loop and can be processed later in case
the current event is still being handled. Removed the session task because
it was just a crude workaround for this problem.

Increased the timeout to 2000ms again since now that time is shared
between the Userport task and the other tasks, another task may well
block a transfer for more than 10ms, for example when WiFi events
are being handled by the WiFi task.

Also added another transfer type for partial sends, since this is required
to first send the response length before actually sending the data.