Command: Added Command base class and echo command, adjusted Service
The Command class is the base class for all commands. It provides a
factory method Command::create(Service::Request) where the concrete
Command class is created based on the Request id and the request is
passed to the Command's constructor.

Commands need to implement the execute() method which returns the
response as a Service::Data instance.

Also implemented a simple echo command which just returns the data
passed in the first request argument for testing purposes.