Tests: Added C64 test program for the echo command
This test program will continually send up to 16kb of random data using
the echo command and validate the response. Written in acme assembler.

The universal routines in universal.asm are based on the official version
113 with only minor improvements to make them easier to include into
other programs. Removed the hardcoded location $C000 and put the code
into it's own zone. The labels in that zone should be localized later.

Also made some minor improvements for readability. The functionality
must remain unchanged in order to maintain compatibility with existing
binaries.