universal.asm: Fixed bogus timeout code, now fail after first timeout
The previous implementation would not abort a transfer if a timeout
occured, it would simply set the timeout to the lowest possible value
(about 500ms) and continue with the transfer anyways. This meant that
if the esp was non-responsive, the user would have to wait for
remainingBytes*500ms until the push or pull function would return, only
to check the timeout and realize the transfer failed. So even when the
user configured a low timeout value, the actual timeout threshold was still
relatively huge.

This also caused distress on the ESP side in case the ESP was reset while
receiving data from the C64, because the C64 would still be happily sending
data which the ESP could only interpret as bogus attempts to initialize
a request (wrong API ID mostly, but every once in a while a $57 (W) by
accident). Always fail early.