Test: Separated echo test, added resistance test and main menu
Tests are now separate from main.asm and reside in individual files
in the tests directory.

The noise resistance test can be used to verify that the firmware
is resistant against random line noise (as produced by timed out
transfers in previous versions of the universal routines). The test
simply transfers random data sprinkled with a few API ID bytes to
the ESP.

The user can choose from the available tests in the main menu.

RESTORE aborts a running test and returns to the menu, while
RUNSTOP/RESTORE aborts the test program.