Firmware: Hardended against line noise, added command registry, namespaced
Userport:

In the calls to esp_event_post_to() the ticksToWait argument has been
set to zero so that events will no longer be posted once the event loop
queue is full. This avoids ISR watchdog timeouts when being confronted
with random line noise. A small vTaskDelay() was added it acceptRequest()
when an unknown API id was received to avoid task watchdog timeouts as
well.

Service:

Fixed a memory leak caused by not deleting the command containing the
argument data when no response data was returned from the command.

Also fail early in parseRequestHeaderV1() when an undefined command id
is requested, avoiding futile creation of a command instance and receipt
of argument data.

Command:

Implemented a command registry that maps the request IDs to the concrete
Command classes, using an std::map of ids to std::function's that create
the corresponding commands. Command::create() has been refactored to use
the command registry instead of creating the Command instances manually.
The static Command::defined() method has been added to be able to tell in
advance whether a mapping exists for a given id. This is used to fail
early when unknown command id is requested in Service.

The Undefined Command has been removed and the property emptyResponse has
been added to the Command class. This property is initialized with an
empty Service::Data instance and is deleted in ~Command(). Thus commands
that return no response can simply return emptyResponse() and need not
to think about freeing the empty Data instance on destruction. The default
execute() implementation returns this instance by default.

The command registry itself has been placed into commands/commands.h and
the macros WIC64_COMMANDS and WIC64_COMMAND have been defined to make
adding new commands straightforward and less error-prone.

In addition all code has been moved to the Wic64 namespace in order to
avoid future name clashes and to avoid placing the command registry into
global scope.

Also the hexdump() utility function has been renamed to log_data().