Userport: Fixed exception on receiving handshake while still in constructor
When a handshake was received immediately after startup, posting the
corresponding event caused a LoadProhibited exception. This happened
because the initial connect() call was done from the Userport constructor,
so it was possible that the ISR was called before the Userport instance
was properly constructed.

Consequently moved the inital connect() call to the setup() function.