Userport: Added onFailureCallback for receiving transfers
Added support for optional failure callbacks that can be passed to
receive() and receivePartial() to allow freeing memory already allocated
for a request if the transfer receiving the request argument times out.

Added typedef callback_t to userport.h to improve readability.

In addition, also check if callbacks are defined before calling them,
because the user might still choose to pass NULL to transfer methods.