Userport: Initialize callback pointers, null failure callback after call
This needs to be done to avoid that, when aborting a transfer due to an
unknown id, a leftover failure callback from a previous transfer is not
called again.

Also uninitialized pointers point to random addresses in C/C++. Need to
check that other pointers are properly initialized to NULL later on.