Data/Service: Allow to optionally specify a different size to report
Added sizeToReport(uint16_t) and sizeToReport(). This is needed for api
version 1 where requests for .prg files are transfered with a reported
size excluding two bytes for the load address. This needs to be addressed
in the future, but for now this quirk gets included for compatibility.

Internally m_sizeToReport is an int32_t initialized with -1. If it is
not set to a value > -1, sizeToReport() returns m_size, else m_sizeToReport.

Adjusted Service to always report sizeToReport() back to the c64.