Userport: Correctly handle final handshake from C64 for sending transfers
The last handshake send from the C64 when receiving data was not handled
correctly before, as I was under the assumption that the FLAG2 line only
gets asserted on writes to $dd01, but of course it also sends a handshake
on read. So when the C64 read the last byte of a transfer, the resulting
final handshake signal was misinterpreted on the ESP side as the intial
handshake for the next transfer.

Added TRANSFER_STATE_TERMINATING for this condition and set it accordingly
in completeTransfer() for sending transfers. Handle it in the handshake
event handler, setting TRANSFER_TYPE_NONE only after the final handshake
has been received.

Reordered completeTransfer(), now call the onSuccessCallback first, then
send the final handshake for all transfers expect RECEIVE_PARTIAL.