Refactored Command => Service interface, implemented queued transfers
Instead of returning the response Data object directly from their execute
methods, Commands now call their responseReady() method, which posts the
corresponding SERVICE_RESPONSE_READY event to the newly added Service
event loop. Responses now come in two flavors, they either contain the
data itself (as before) or they contain an RTOS queue.

When the Service component receives the reponse ready event, it calls
Data::isQueued() to determine whether it is a queued or a static response.
Static responses are send down in one transfer, while queued responses
are send down in chunks of 32 bytes in multiple partial send transfers.

The client now uses this mechanism to start sending data as soon as it
starts to read the response from the http connection, instead of always
downloading the whole response first. This works only for http responses
that contain a content length header, since the response size must be known
in advance to the C64. For chunked transfers that don't supply the content
length up front the data is still only send after the client has read it
in its entirety.

Nevertheless, in theory this allows for the transfer of payloads that
exceed the size of the ESPs and/or the C64's memory. Currently though
transfers to the C64 are limited to 64kb by the data transfer protocol
(only two bytes are sent for size). Also streaming indefinite amounts
of data could be implemented in this fashion.

Also improved error handling and managing keep-alive connections in the
Client, properly reusing the existing connection if keep-alive is requested
by the server.

Also adjusted the other commands to use the new event mechanism and
generally improved code.