Client: Added handling of WiC64-Security-Token-Key header
This is an alternative to checking if the status code is 201 and then
sending the security token key in a custom format in the response body.

This header can simply be added to the servers sid.php so that the old
firmware will still get the expected data while this firmware will just
get the key from the http response header. Since the headers are already
parsed by the esp_http_client, the firmware does not need to implement
parsing a custom format.

Also made some minor improvements to variable names and formatting.