WiC64: Changed memory allocation scheme, added webserver component
No longer malloc() memory for the request argument, instead use the
same global 64kb transfer buffer for both the response and the request.
Since the both are limited to 64kb, a single buffer will do. This saves
memory, avoids repeated mallocs and thus reduces the possibility of heap
fragmentation.

This implies that once a command starts filling the response, the previous
request data will be overwritten, so a command that needs to analyze or
adjust the request data must do so before setting its response data.

The Data class has been adjusted accordingly, the Data(size) constructor
has been removed. Data now simply wraps pointer and size, and allows
getting the data as a c-string via the added c_str() method (the global
buffer  actually contains 64kb + 1 byte to ensure c_str() always returns a null-
terminated string).

If we add commands with multiple arguments later, which is already
prepared in the way Request is implemented, we will have to introduce
a pointer to the next free region in the buffer. Will keep it simple
and won't reinvent malloc(), just initially divide the buffer between
arguments when building the request.

In addition, the size of the reponse queue has been reduced to 16kb, which
seems to be enough, at least for now.

Also added a simple webserver component which allows the user to switch
the debug level at runtime, e.g. http://192.168.0.64/?log=DEBUG.