Redesigned and stabilized interrupt based transfers and protocol handling
Switched back to using an interrupt based approach. All low level
communication is now handled in the handshake interrupt while the high
level events are handled from the Userport and Service event loop tasks.
The Userport now listens to the REQUEST_ACCEPTED, TRANSFER_COMPLETED and
READY_TO_SEND events posted from the handshake ISR.

The data direction interrupt has been removed, since it is not necessary
at all to react to changes on this line to perform the transfer. It is
obvious from the protocol when this line should change, we now merely
read the line to make sure the C64 is in the proper state before changing
the esp data direction and start receiving or sending.

Once the request has been successfully received, the C64 is waiting for
the next handshake, so once the response is ready to send, we initiate
the send by sending the handshake from startTransfer() and setting the
transfer state to PENDING. The C64 then sends another handshake, upon which
the ISR posts READY_TO_SEND. The transfer is then initiated by switching
the port to output and writing the first byte from onReadyToSend().

This is much simpler than having to deal with two different interrupt
sources and making sure one interrupt is not missed while handling the
other.

An interrupt based approach also boosts performance considerably, as
compared to posting an event on every handshake. Transfer speeds were
substantially improved by this measure.

Removed all logging from ISRs because the esp log functions block and
thus starve the ISR watchdog. Also added vTaskDelay(1) functions to yield
to other threads from Service and the Clients queueing task. Reenabled
both the ISR and task watchdogs.

Fixed a serious bug in the Client that caused queued transfers to fail
sporadically. The bug was caused by passing a pointer to the local
variable content_length from Client::get() to the queue task, which only
works if the stack location pointed to has not been reused by chance in
the meantime. Unfortunately, this did work most of the time, so it took
me a while to find it.