Data: Initialize m_data with transferBuffer, added copyFrom(), fixes
Renamed the data()-Methods to wrap(), since they wrap data allocated
elsewhere.

Added copyFrom() to actually copy data into the transferBuffer. Also
intialize the m_data member with a pointer to the global transferBuffer.

The copyFrom() method can be used when the response data is created
dynamically somewhere else and is deallocated later, for example when
using a local  String instance to build a response.

Also "fixed" wrap(const char* c_str()) and don't add 1 to m_size for
the terminating nullbyte. This needs to be done for compatibility with
the original firmware, where ALL data was send from arduino String
instances and the response size was always just set to String.length(),
even when the reponse actually constituted a single string and a
terminating null byte would have come in handy for the receiver as well.

Why this has been done this way is beyond me, since null-terminating
strings is a common practice on the C64 as well, especially when using
assembly.