Command: Use simple struct instead of std::map for the command map
There is a bug in ESP-IDF 4.x causing std::map::find() to throw
exceptions in some rare edge cases, presumably fixed in 5.x, but
the arduino-esp32 component can not yet be used with ESP-IDF >= 5.

commands/commands.h remains unchanged, just adjusted the macros in
command.h and the implementations of Command::defined() and
Command::create().