Data: added addSeparated() methods to add separated values
These methods copy the contents of a string to the buffer pointed to
by Data::m_data at the index maintained in m_index and also append a
separator char, adjusting m_index and m_size accordingly. If no separator
is specified, the default separator 0x01 is used.

This makes it easier to pass data in the 0x01-separated format introduced
by the existing firmware.

Scan: Adjusted to make use of Data::addSeparated()