Commands: Added connect command, fixed reconnect behaviour and scanning
Implemented command Connect (0x0d = 13), i.e. set WiFi credentials based
on the index of the network in the last scan result and the passphrase
supplied by the user.

For this is was necessary to refactor Connection and the Scan command,
since scanning can not be done while connection attempts are being made,
and we formerly always tried to reconnect when receiving a disconnect
event. It was somewhat pointless to keep trying to reconnect forever if
the connection kept failing due to wrong credentials as well.

Thus we now actively manage the automatic reconnect in Connection in the
member variable m_reconnecting, and only set it to true once a connection
is actually established. The disconnect() method now disconnects and
sets m_reconnecting to false, so it's an explicit request to disconnect
and *not* try to reconnect until a successive connect() call results in
an actual connection.

Implemented scanNetworks() in Connection to first disconnect(), call
WiFi.scanNetworks() and then connect() again to make sure scanning for
networks always works. There seem to be some quirks related to scanning
and multiple contradicting workarounds are suggested here:

https://github.com/espressif/arduino-esp32/issues/3294

These workarounds involve manual delays, which seems a bit voodoo, so
I have decided that simply disconnecting before a scan and reconnecting
again once it is done is the most pragmatic approach here.

Scan: Now use connection->scanNetworks() instead of directly calling
WiFi.scanNetworks(). Also added responding with "no networks found"
like the previous firmware does, additionally logging whether the scan
failed in the first place or whether it succeeded and there are really
no networks available (see comment).

Data: Added method field(index) to access 0x01-separated fields. This is
the equivalent to the getValue()-function in th previous firmware.
Also renamed appendSeparated() to appendField().