Connect: Add a 3000ms delay before sending response (compatibility)
After receiving the response, the portal's wifi.prg (and probably
launcher.prg) immediately try to determine whether a connection has been
established by requesting the ip nine times in quick succession before
giving up and printing "Not connected. Wrong password?".

The previous firmware added a delay of 3 seconds so that the client would
not start checking the IP right away.

Though I can see why this makes the clients life easier, it is not the
firmware's job, at least not when doing it implicitly while client code
is silently relying on this behaviour. An alternative would be to
introduce an explicit command to check for a connection that allows the
user to specify a timeout. This would also simplify client code further.