Connect: Implemented decoding of escaped ASCII characters in passphrase
Since some ASCII characters can not be entered on the C64, these
characters can be escaped in the WiFi passphrase entry using a PETSCII
up-arrow character (ASCII tilde), followed by two hex digits, e.g.
"~5f" escapes an ASCII underscore character (0x5f).

Implemented decoding with proper error handling/reporting, including
rejection of passphrases that exceed the maximum length of 63 characters
for WPA passphrases.

In general, if the passphrase is invalid or the SSID is empty, an
appropriate error message is returned in the response, and no further
action is taken. This deviates from the behaviour of the previous
firmware, where an implicit reconnect using the previously stored
credentials was performed (WiFi.begin()) if the ssid was empty
(the passpharse decoding did not implement error checking at all).