Service: finalizeRequest(): prevent NPE if request already finalized
Some portal programs, notably the binary and "berlin" clock programs
send additional handshakes during change of transfer direction when
requesting the IP address. This causes the pending send transfer to be
initiated twice and ultimately results in finalizeRequest() being called
as well. As a first measure, prevent the resulting NPE and log an error
message instead, hinting at a possible violation of protocol or a firmware
bug. Need to evaluate whether this condition can be guarded against.

The misbehaving programs are BASIC programs that seem to contain a very
early version of the universal routines, as the REM comments state
a date Tag of 10-2021.