Userport: Added delays between consecutive handshakes during start/complete
When starting and completing transfers, small delays are needed before the
initializing and/or final handshakes, especially in startTransfer(), where
two consecutive handshakes are send in case of starting a partial send
transfer. Without a delay, the C64 may not get around to clear the handshake
flag in $dd0d before we sent the second handshake. A delay of 10ms seems
to be enough to avoid this.

This problem becomes apparent if the C64 sends commands in quick
succession. Removed the artificial delay in the wifi info test that was
added in the last commit to temporarily work around this issue.