Userport: Hopefully fixed sporadically missed handshake interrupts
There was a nasty bug that would sporadically cause transfers to time out,
even when thousands of transfers succeeded before. The handshake interrupt
would occasionally be missed by the ESP. This seems to be due to a problem
with edge triggered interrupts in the ESP GPIO, as described here:

https://www.quantulum.co.uk/blog/esp32-edge-triggered-interrupt-bug/

Although the described problem does not fully apply here because we were
using just one edge triggered interrupt, it seems that switching to a
level interrupt solves the problem, although more long duration tests are
needed.

I also simply copied the attachInterrupt calls from the original firmware,
where the bogus Arduino interrupt type value HIGH was used instead of
ONHIGH, which accidentally resulted in actually using a rising edge
trigger (Arduino does not use a proper enum type here, so the compiler
won't catch such unintended input). However, I also didn't notice that the
PC line is actually active low, so I just changed it to RISING, which sort
of worked most of the time, apparently.

I've now used the esp API to set up the interrupt as a low level interrupt.
The workaround described in the link above is not necessary here because
the PC line stays low only for one cycle, so switching to a a high level
triggered interrupt every other time is not required here, the ISR always
takes more than 1us.

Since debugging this was quite a pain I added some more stuff that might
turn out to be just snake oil, but I will leave it in for now and just
phase that out over time. Testing for this bug takes a lot of time since
the bug might not surface for several hours at a time.

I did remove the pullups and pulldowns from the handshake and data direction
lines since they are already pulled up on the C64, and turning off the C64
while the ESP is still powered over USB leads to bogus handshakes being
detected.