Userport: Implemented premature abortion of transfers for compatibility
In the existing protocol it is possible for the client to issue a new
request even though the previous response has not been completely
received. This is possible because the original firmware resets its
transfer state whenever the data direction line goes high, i.e.
interpreting a positive edge on PA2 as an initial signal for a request
regardless of the state of any currently running transfer.

For example, the current transfer routines for loading the portal simply
abort and retry the request immediately if the first byte of the payload
happens to be $21 (ascii '!') (http get error response is "!0") without
receiving the complete response payload first. Aside from unecessarily
complicating the protocol, this also introduces the possibility of an
infinite loop as long as the portal code can not be retrieved from the
server for some reason, including intermittent loss of wifi connectivity
or server downtime.

Due to the lack of proper handshaking in the initializing phase of the
current protocol the PA2 line is thus still required unless the
communication protocol is changed to a more common approach, where
handshaking is used to initiate a transfer and to ensure mutual readiness
for communication first.

In order to implement this behaviour an interrupt on the positive edge
of the data direction line PA2 has been added again, resetting the current
transfer state to TRANSFER_STATE_NONE. In addition, a delay of 1ms is
required in onTransferCompleted() to allow the transition from
TRANSFER_STATE_TERMINATING to TRANSFER_STATE_NONE to occur.