wic64.asm: Added option wic64_optimize_for_size, improved comments
Added assembly time option "wic64_optimize_for_size". If set to a non-zero
value, a subroutine will be used to wait for handshakes instead of a
macro. This will bring down the size from 568 to 322 bytes, but will also
decrease transfer speeds by about 30%. Intended to be used when size is
more important than speed, for example when transfering high scores.

Also added an overloaded macro for wic64_execute without an explicit
timeout argument. The timeout will be set to the default $02 when this
macro is used.

Adjusted the existing test cases accordingly.