TcpClient, Tcp: Implemented TCP commands open, read, write
Tcp commands open (0x21), read (0x22) and write (0x23) have been
implemented in the same manner as in the previous implementation.

The telnet demo program available in the portal should thus work
as before.

Data: Added setter for size which automatically adds a nullbyte after
the last byte as a side-effect to ensure the data can be treated as
a null-terminated string, added method zeroTerminated() which adds a
nullbyte to the actual data (also increasing the data size by one).