Userport/Display: Don't show userport state when connecting/disconnecting
It makes no sense to display the user port connection state to the user
all of the time, since the userport will only be disconnected when the
user explicitly chooses to do so by pressing the Wic64-Button
(not implemented yet).

In this case it makes more sense to display a short notification to the
user to confirm that the wic is in the desired state.