Tests: Removed noise resistance test
This test is futile, since the firmware can not be completely hardened
against line noise. Random data may always include patterns that are
interpreted as seemingly valid commands. Dealing with all kinds of bogus
input would make the code unnecessarily complex.

If a stray client program misbehaves in this way, it won't work properly
anyways and the user is likely to abort it. A hard reset is ultimately
the right behaviour for the esp in these cases.