wic64.asm: Removed option wic64_blank_screen
This is not really necessary, blanking the screen is for the user to
decide. wic64_enable_irqs is different in this regard, because disabling
the irqs should be the default for speed/stability unless the user
explicitly requests a different behaviour (in which case we assume the
user knows what he's doing).