Buttons: Implemented button functions, improved Display and Settings
ESP Boot button: disconnect/connect userport

Displays a permantent message when the userport is disconnected and the
WiC64 is effectively disabled. No need to disconnect the WiFi or even
restart the ESP as it was done in the original firmware.

Wic64 Button: short press => Enable/disable LED

Displays a short notification about the resulting LED state.

Wic64 Button: long press => Rotate display by 180 degrees

No feedback as the result is apparent on the display.

These options are persisted and the WiC64 will start up in the configured
state on next boot. Added the necessary methods to Settings, using the
same keys as in the orginal firmware.

Updated TODOs, removed TODOs for unused UDP commands