Version: Added WIC64_VERSION_DEVEL (commits since last stable version tag)
Added a fourth version number denoting the commits since the last official
version derived from the corresponding git tag.

For versions that match the official release, this number will be set to
zero. Any nonzero value thus denotes an unstable development version. This
will allow the user to upgrade to a development version (given that we
publish it on the server) in case this is relevant for collaboration with
third party developers.

The version number command ($26) WIC64_CMD_VERSION_NUMBERS now accordingly
responds with four bytes <major>, <minor>, <patch> and <devel>.

The display shows the WIC64_SHORT_VERSION string, which is either

<major>.<minor>.<patch> => stable "official" versions

or

<major>.<minor>.<patch>-<devel> for unstable development versions

The firmware update will allow the user to upgrade to an unstable version
if it is published on the server.